requirements.txt
# XScanner Dependencies

# Core dependencies
asyncio
aiohttp>=3.8.0
pyyaml>=6.0
pandas>=1.5.0
python-dateutil>=2.8.0

# Twitter scraping
snscrape>=0.7.0
tweepy>=4.12.0  # For X API v2 (optional)

# Data processing
numpy>=1.24.0
jsonschema>=4.17.0

# Storage and export
pymongo>=4.3.0  # For MongoDB (optional)
notion-client>=2.0.0  # For Notion integration (optional)

# HTTP requests and API calls
requests>=2.28.0
httpx>=0.24.0

# Utilities
tqdm>=4.65.0  # Progress bars
python-dotenv>=1.0.0  # Environment variables
schedule>=1.2.0  # Task scheduling

# Future integrations
python-telegram-bot>=20.0  # Telegram notifications (optional)
google-api-python-client>=2.0.0  # Google Sheets (optional)

# Development and testing (optional)
pytest>=7.2.0
pytest-asyncio>=0.21.0
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Code linting